requirements.txt
opencv-python
onnxruntime-gpu
insightface
imageio
pillow
